% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{add_geom_group_id}
\alias{add_geom_group_id}
\title{Ajouter un identifiant basé sur la géométrie}
\usage{
add_geom_group_id(x, col_name = "geom_id")
}
\arguments{
\item{x}{`sf`, objet spatial}

\item{col_name}{Character, nom de la colonne à créer contenant l'identifiant
(default: "geom_id")}
}
\value{
Un objet `sf` avec une colonne supplémentaire identifiant les groupes
de géométries identiques
}
\description{
Génère un identifiant pour chaque géométrie d'un objet `sf` afin de repérer
les géométries dupliquées (géométries strictement identiques).
}
