% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{assemble_q_canonised}
\alias{assemble_q_canonised}
\title{Assemblage des données canonisées}
\usage{
assemble_q_canonised(q_multi_r_multi, q_multi_r_uni, q_uni)
}
\arguments{
\item{q_multi_r_multi}{`data.frame`, questions à choix multiples réponses multiples}

\item{q_multi_r_uni}{`data.frame`, questions à choix multiples réponses uniques}

\item{q_uni}{`data.frame`, questions sans choix}
}
\value{
Un `data.frame` final canonisé
}
\description{
Combine les différentes catégories de questions (choix multiples, uniques,
sans choix) en un seul jeu de données harmonisé.
}
