% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseaux.R
\name{assign_id_reseau_to_mares}
\alias{assign_id_reseau_to_mares}
\title{Associer les mares à l'identifiant de leur réseau}
\usage{
assign_id_reseau_to_mares(kobo, reseau)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{reseau}{`sf`, couche de polygones représentant les réseaux de mares}
}
\value{
Un objet `sf` avec les identifiants des reseaux de mares.
}
\description{
Pour chaque mare, ajoute l'identifiant du réseau polygonal auquel elle appartient
ainsi que le nombre de mares dans ce réseau.
}
