% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_iecmar.R
\name{calcul_iecmar}
\alias{calcul_iecmar}
\title{Calcul des scores IECMAR}
\usage{
calcul_iecmar(df)
}
\arguments{
\item{df}{Un data.frame contenant au minimum les colonnes `X_uuid` et `cor_iecmar`.}
}
\value{
Un data.frame enrichi avec les colonnes `points`, `n` et`note`.
}
\description{
Calcule les notes IECMAR à partir d’un jeu de données de réponses.
La fonction joint les points associés, applique un traitement des doublons,
puis calcule une note brute et une note pondérée sur 20 par individu (`X_uuid`).
}
