% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_iecmar.R
\name{calcul_mediane_iecmar_reseaux}
\alias{calcul_mediane_iecmar_reseaux}
\title{Calcul de la médiane IECMAR par réseau}
\usage{
calcul_mediane_iecmar_reseaux(df)
}
\arguments{
\item{df}{`data.frame`, contenant au moins les colonnes :
\code{id_reseau}, \code{note}}
}
\value{
Un `data.frame` enrichi avec :
\describe{
  \item{median_iecmar_reseau}{Médiane des notes brutes par réseau}
  \item{median_iecmar_reseau_pon}{Médiane des notes pondérées par réseau}
  \item{position_mediane}{Indique si la note est au-dessus ou en dessous de la médiane}
}
}
\description{
Pour chaque réseau identifié par `id_reseau`, calcule la médiane
des notes brutes et pondérées, et indique la position de chaque mare
par rapport à cette médiane.
}
