% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{calculs_dechets}
\alias{calculs_dechets}
\title{Notation des déchets et pollution (codes C18 et C19) - V4 uniquement}
\usage{
calculs_dechets(cano)
}
\arguments{
\item{cano}{`data.frame`, données mares du kobotoolbox post-canonisation}
}
\value{
Le `data.frame` original avec les lignes correspondants aux dechets et quantites
dechêts modifiées pour correspondre à la notation iecmar (dechêts & pollution)
}
\description{
Pour chaque mare, évalue la présence de déchets et leur quantité, puis évalue la pollution.
}
