% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{calculs_habitats_v4}
\alias{calculs_habitats_v4}
\title{Calcul des codes IECMAR pour le type de mare et les habitats (V4)}
\usage{
calculs_habitats_v4(canonique)
}
\arguments{
\item{canonique}{`data.frame`, jeu de données canonisé avec colonnes
`X_uuid`, `CAN_name` et `CAN_choice`}
}
\value{
Un `data.frame` contenant toutes les lignes originales ainsi que les lignes
calculées pour `type_mare` avec la colonne `cor_iecmar` indiquant le code
}
\description{
Pour chaque mare, détermine le type de mare et les habitats associés, puis
attribue le code IECMAR correspondant selon les règles définies.
}
