% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{calculs_helophytes}
\alias{calculs_helophytes}
\title{Calcul du recouvrement en hélophytes et codage associé}
\usage{
calculs_helophytes(cano)
}
\arguments{
\item{cano}{`data.frame`, données canonisées}
}
\value{
Un `data.frame` avec les valeurs transformées et codées
}
\description{
Transforme les valeurs de recouvrement des hélophytes (id_can 210)
en classes selon des seuils définis.
}
