% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{calculs_hydrophytes}
\alias{calculs_hydrophytes}
\title{Calcul du recouvrement en hydrophytes et codage associé}
\usage{
calculs_hydrophytes(cano)
}
\arguments{
\item{cano}{`data.frame`, données canonisées}
}
\value{
Un `data.frame` avec les valeurs agrégées et codées
}
\description{
Agrège les valeurs de recouvrement des hydrophytes (id_can 208 et 209),
puis attribue une classe selon des seuils définis.
}
