% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{calculs_superficie}
\alias{calculs_superficie}
\title{Calcul de la superficie des mares et codage associé}
\usage{
calculs_superficie(cano)
}
\arguments{
\item{cano}{`data.frame`, données canonisées}
}
\value{
Un `data.frame` enrichi avec une ligne calculée de superficie et un code associé
}
\description{
Calcule la superficie des mares à partir de deux dimensions (id_can 200 et 201),
puis attribue une classe de superficie selon des seuils définis.
}
