% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clc}
\alias{clc}
\title{Occupation du sol Corine Land Cover 2018 niveau 1}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 792 rows and 1 columns.
}
\source{
Données originales provenant du Corine Land Cover 2018.
}
\usage{
clc
}
\description{
Jeu de donnees du CLC au format `sf` sur la region Bourgogne-Franche-Comte.
Les geometries ont ete simplifees.
Il est utilisé (facultativement) dans le cadre du calcul des réseaux de mares.
}
\keyword{datasets}
