% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{compile_iecmar_inputs}
\alias{compile_iecmar_inputs}
\title{Compilation des données IECMAR}
\usage{
compile_iecmar_inputs(res_forms, res_geom)
}
\arguments{
\item{res_forms}{Résultats issus des formulaires Kobo.}

\item{res_geom}{Résultats issus des traitements géométriques.}
}
\value{
Un tableau long prêt pour le calcul des notes IECMAR.
}
\description{
Fusionne les résultats issus des traitements géométriques
et des formulaires Kobo dans un format commun.
}
\keyword{internal}
