% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{compute_q_uni}
\alias{compute_q_uni}
\title{Traitement des questions sans choix}
\usage{
compute_q_uni(q_multi_r_uni_before_filter, cor, canonique)
}
\arguments{
\item{q_multi_r_uni_before_filter}{`data.frame`, données préparées en amont}

\item{cor}{`data.frame`, table de correspondance}

\item{canonique}{`data.frame`, table canonique}
}
\value{
Un `data.frame` des questions sans choix
}
\description{
Identifie les questions sans choix (sans correspondance canonique),
sépare les cas avec et sans valeur, puis recombine les résultats.
}
