% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseaux.R
\name{create_raw_reseau}
\alias{create_raw_reseau}
\title{Creation d'un reseau brut de plans d'eau}
\usage{
create_raw_reseau(mares_kobo, eau_surfacique_ign, buffer = 1000)
}
\arguments{
\item{mares_kobo}{`sf`, couche spatiale des mares (ex: issue de Kobo)}

\item{eau_surfacique_ign}{`sf`, couche spatiale des surfaces en eau (ex: IGN)}

\item{buffer}{Numeric, distance en metres appliquee comme tampon autour des geometries
avant fusion (default: 1000)}
}
\value{
Une couche `sf` de polygones representant le reseau brut fusionne
}
\description{
Genere une couche polygonale continue en fusionnant deux couches d'entrees
(mares et plans d'eau), apres application d'un tampon. Cette etape permet
de constituer un reseau initial avant tout decoupage ou fragmentation ulterieure.
}
