% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseaux.R
\name{cut_reseau_by}
\alias{cut_reseau_by}
\title{Decoupage d'un reseau par une couche de fragmentation}
\usage{
cut_reseau_by(reseau, by)
}
\arguments{
\item{reseau}{`sf`, couche polygonale representant le reseau a decouper}

\item{by}{`sf`, couche polygonale utilisee pour le decoupage (ex: routes bufferisees)}
}
\value{
Une couche `sf` de polygones correspondant au reseau decoupe
}
\description{
Coupe un reseau polygonal (ex: mares et plans d'eau) a l'aide d'une couche
de polygones representant des elements fragmentants (prealablement buffes).
Cette operation permet de segmenter le reseau en entites distinctes.
}
