% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eau}
\alias{eau}
\title{Surfaces d'eau de la BD TOPO de l'IGN}
\format{
An object of class \code{list} of length 8.
}
\source{
Données originales provenant de la BD TOPO IGN (2025).
}
\usage{
eau
}
\description{
Ce jeu de données contient les polygones représentant les petites surfaces
d'eau (mares, plans d'eau). Les geometries ont ete simplifees.
Il est utilisé dans le cadre des traitements géométriques pour le calcul du IECMAr.
}
\keyword{datasets}
