% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eau_max1ha}
\alias{eau_max1ha}
\title{Surfaces d'eau de la BD TOPO de l'IGN ≤ 1000 m²}
\format{
An object of class \code{list} of length 8.
}
\source{
Données originales provenant de la BD TOPO IGN (2025).
}
\usage{
eau_max1ha
}
\description{
Ce jeu de données contient les polygones représentant les petites surfaces
d'eau (mares, plans d'eau) filtrées à 1000 m² maximum.
Les geometries ont ete simplifees.
Il est utilisé dans le cadre du calcul des réseaux de mares.
}
\keyword{datasets}
