% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{finalize_q_multi_r_uni}
\alias{finalize_q_multi_r_uni}
\title{Finalisation des questions à choix multiples – réponses uniques}
\usage{
finalize_q_multi_r_uni(q_multi_r_uni_before_filter, q_uni)
}
\arguments{
\item{q_multi_r_uni_before_filter}{`data.frame`, données préparées}

\item{q_uni}{`data.frame`, questions sans choix}
}
\value{
Un `data.frame` filtré
}
\description{
Exclut les observations correspondant aux questions sans choix afin de ne
conserver que les questions à choix multiples avec réponse unique.
}
