% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forets}
\alias{forets}
\title{Forêts (entitées ligneuses de l'OCS GE) > 1 ha}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 47853 rows and 2 columns.
}
\source{
Données originales provenant de l'OCS-GE 2023
}
\usage{
forets
}
\description{
Couche de l'OCS-GE, filtré pour ne garder que les forêts (au sens large) de plus de 1 hectare.
Les geometries ont ete simplifees.
Il est utilisé dans le cadre des traitements géométriques pour le calcul du IECMAr.
}
\keyword{datasets}
