% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{iecmar_c11}
\alias{iecmar_c11}
\title{Distance à la pièce d’eau la plus proche (C11) : Calcul de la distance minimale aux mares et plans d'eau}
\usage{
iecmar_c11(kobo, surface_eau)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{surface_eau}{`sf`, couche polygonale représentant les pièces d'eau}
}
\value{
Un `data.frame` contenant :
\describe{
  \item{X_uuid}{Identifiant de la mare}
  \item{c11}{Code C11 attribué selon la distance minimale aux autres mares ou
  aux plans d'eau (31, 32 ou 33)}
}
}
\description{
Pour chaque mare, calcule la distance minimale soit à une autre mare,
soit à une pièce d'eau de référence, puis attribue la note IECMAr (C11)
en fonction de cette distance.
}
