% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{iecmar_c12}
\alias{iecmar_c12}
\title{Nb de pièces d’eau à moins de 500 m (C12) : Calcul du nombre de mares + pièces d'eau à proximité}
\usage{
iecmar_c12(kobo, surface_eau)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{surface_eau}{`sf`, couche polygonale représentant les pièces d'eau}
}
\value{
Un `data.frame` contenant :
\describe{
  \item{X_uuid}{Identifiant de la mare}
  \item{c12}{Code C12 attribué selon le nombre de pièces d'eau à proximité
  (34, 35 ou 36)}
}
}
\description{
Pour chaque mare, calcule le nombre de pièces d'eau situées dans un rayon
de 500 mètres (en utilisant les centroïdes) puis attribue la note IECMAr (C12)
en fonction de ce nombre.
}
