% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{iecmar_c13}
\alias{iecmar_c13}
\title{Distance avec un site terrestre hivernal (C13) : Calcul de la distance aux forêts}
\usage{
iecmar_c13(kobo, site_hiver)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{site_hiver}{`sf`, couche de points ou polygones représentant les sites
terrestres hivernaux}
}
\value{
Un `data.frame` contenant :
\describe{
  \item{X_uuid}{Identifiant de la mare}
  \item{c13}{Code C13 attribué selon la distance au site hivernal (37, 38 ou 39)}
}
}
\description{
Pour chaque mare, calcule la distance à un site terrestre hivernal et
attribue la note IECMAr (C13) selon la distance (en mètres).
}
