% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{iecmar_c15}
\alias{iecmar_c15}
\title{Zone d’écrasement d’Amphibiens (C15) : Calcul de la distance aux routes}
\usage{
iecmar_c15(kobo, routes, buffer = 250)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{routes}{`sf`, couche linéaire représentant les routes avec colonnes
`cpx_classe` et `nature`}

\item{buffer}{Numeric, distance en mètres pour définir le voisinage autour des mares (default: 250)}
}
\value{
Un `data.frame` contenant :
\describe{
  \item{X_uuid}{Identifiant de la mare}
  \item{c15}{Code C15 attribué selon la proximité et le type de route (43, 44 ou 45)}
}
}
\description{
Pour chaque mare, évalue la proximité des routes et attribue la note IECMAr (C15)
selon la classe et la nature de la route.
}
