% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{join_q_multi_r_multi}
\alias{join_q_multi_r_multi}
\title{Jointure des questions à choix multiples – réponses multiples}
\usage{
join_q_multi_r_multi(kobo, cor, canonique)
}
\arguments{
\item{kobo}{`data.frame`, données Kobo au format long}

\item{cor}{`data.frame`, table de correspondance entre variables Kobo et choix}

\item{canonique}{`data.frame`, table de correspondance canonique}
}
\value{
Un `data.frame` enrichi avec les informations de correspondance
}
\description{
Filtre les variables correspondant à des questions à choix multiples avec
réponses multiples, transforme les modalités sélectionnées et réalise la
jointure avec les tables de correspondance et canonique.
}
