% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgv}
\alias{lgv}
\title{Linéaire de la Ligne a Grande Vitesse (LGV) de la BD TOPO de l'IGN}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 1 columns.
}
\source{
Données originales provenant de la BD TOPO IGN (2025).
}
\usage{
lgv
}
\description{
Jeu de donnees de la LGV au format `sf` sur la region Bourgogne-Franche-Comte.
Il est utilisé dans le cadre du calcul des réseaux de mares.
}
\keyword{datasets}
