% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{minimal_distance}
\alias{minimal_distance}
\title{Calcul de la distance à l'entité la plus proche}
\usage{
minimal_distance(a, b)
}
\arguments{
\item{a}{`sf`, couche spatiale de référence (ex: points)}

\item{b}{`sf`, couche spatiale cible pour le calcul de distance (ex: polygones)}
}
\value{
Un objet `sf` identique à `a` avec une colonne supplémentaire
`dist_surfacique` correspondant à la distance à l'entité la plus proche de `b`
}
\description{
Calcule, pour chaque entité d'un objet `sf`, la distance à l'entité la plus
proche dans une seconde couche spatiale.
}
