% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{nb_entites_dans_rayon_centroids}
\alias{nb_entites_dans_rayon_centroids}
\title{Calcul du nombre d'entités dans un rayon donné autour de chaque point}
\usage{
nb_entites_dans_rayon_centroids(a, b, rayon = 500)
}
\arguments{
\item{a}{`sf`, couche de points de référence}

\item{b}{`sf`, couche de polygones dont les centroïdes seront considérés}

\item{rayon}{Numeric, distance en mètres pour définir le voisinage (default: 500)}
}
\value{
Un objet `sf` identique à `a` avec une colonne supplémentaire
`n_within_500m` indiquant le nombre d'entités dans le rayon autour de chaque point
}
\description{
Pour chaque entité d'une couche de points, compte le nombre d'autres entités
(points ou centroïdes de polygones) situées dans un rayon spécifié.
}
