% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{prepare_data_foret}
\alias{prepare_data_foret}
\title{Préparation de la couche forêt en excluant les zones autour des mares}
\usage{
prepare_data_foret(kobo, forets, tampon = 100)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{forets}{`sf`, couche de polygones des forêts > 1ha provenant de l'OCS-GE, simplifié
à 95%}

\item{tampon}{Numeric, distance en mètres utilisée pour créer un tampon autour
des mares (default: 100)}
}
\value{
Une couche `sf` correspondant aux surfaces forestières excluant les
zones tampon autour des mares
}
\description{
En réponse aux contraintes de la C13 du IECMAr : "Distance avec un site terrestre hivernal (pas pris
en compte à moins de 100m)", supprime des surfaces forestières les zones situées à proximité des
mares en appliquant un tampon autour de celles-ci.
}
