% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{prepare_data_hydro}
\alias{prepare_data_hydro}
\title{Préparation de la couche hydrographique en excluant les zones proches des mares}
\usage{
prepare_data_hydro(kobo, hydro, tampon = 10)
}
\arguments{
\item{kobo}{`sf`, couche de points représentant les mares}

\item{hydro}{`sf`, couche polygonale des surfaces en eau}

\item{tampon}{Numeric, distance en mètres utilisée pour exclure les surfaces
en eau proches des mares (default: 10)}
}
\value{
Une couche `sf` filtrée des surfaces en eau
}
\description{
Filtre les surfaces en eau pour exclure celles situées à proximité des mares
déjà étudiées, en appliquant un tampon autour des points Kobo.
}
