% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Préparation des données d'entrée}
\usage{
prepare_input_data(df, version, departement, use_history_if_exist, jdd_v4)
}
\arguments{
\item{df}{Objet sf contenant les données de mares.}

\item{version}{Version du formulaire Kobo utilisée.}

\item{departement}{Code du département.}

\item{use_history_if_exist}{Booléen d'utilisation de l'historique.}

\item{jdd_v4}{Liste des jeux de données V4.}
}
\value{
Une liste contenant les jeux de données préparés et les
informations de versionnement.
}
\description{
Filtre les mares existantes et ajoute les données
historiques lorsque celles-ci sont disponibles.
}
\keyword{internal}
