% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{prepare_negatif_layer}
\alias{prepare_negatif_layer}
\title{Preparation d'une couche lineaire pour decoupage de reseau (utilise dans data-raw et reseaux (option routes RD))}
\usage{
prepare_negatif_layer(layer, buffer, nQuadSegs = 5)
}
\arguments{
\item{layer}{`sf`, couche spatiale lineaire a transformer}

\item{buffer}{Numeric, distance en metres pour la creation des tampons}

\item{nQuadSegs}{Integer, nombre de segments utilises pour approximer les courbes
lors du buffer (default: 5)}
}
\value{
Une couche `sf` de polygones valides issue de la couche lineaire d'entree
}
\description{
Transforme une couche lineaire en polygones via l'application de tampons etroits,
afin de permettre le decoupage ulterieur de reseaux polygonaux. Cette fonction est
typiquement utilisee pour representer des elements fragmentants (ex: routes, voies ferrees).
}
