% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{prepare_q_multi_r_uni}
\alias{prepare_q_multi_r_uni}
\title{Préparation des questions à choix multiples – réponses uniques}
\usage{
prepare_q_multi_r_uni(kobo, cor, canonique)
}
\arguments{
\item{kobo}{`data.frame`, données Kobo au format long}

\item{cor}{`data.frame`, table de correspondance}

\item{canonique}{`data.frame`, table canonique}
}
\value{
Un `data.frame` avec les jointures appliquées
}
\description{
Sélectionne les variables correspondant à des questions à choix multiples
avec réponse unique et applique les jointures de correspondance.
}
