% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_iecmar_forms}
\alias{process_iecmar_forms}
\title{Traitement des formulaires Kobo}
\usage{
process_iecmar_forms(df, version, multiple_form_v)
}
\arguments{
\item{df}{Objet sf contenant les données de mares.}

\item{version}{Version du formulaire Kobo utilisée.}

\item{multiple_form_v}{Booléen indiquant la présence de plusieurs versions.}
}
\value{
Un tableau long contenant les variables IECMAR issues des formulaires.
}
\description{
Convertit les réponses Kobo vers le format canonique
IECMAR et applique les calculs spécifiques aux versions
de formulaire.
}
\keyword{internal}
