% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_iecmar_geom}
\alias{process_iecmar_geom}
\title{Traitements géométriques IECMAR}
\usage{
process_iecmar_geom(df, departement)
}
\arguments{
\item{df}{Objet sf contenant les données de mares.}

\item{departement}{Code du département.}
}
\value{
Un tableau long contenant les résultats des traitements géométriques.
}
\description{
Réalise les calculs géométriques nécessaires à la
production des indicateurs IECMAR.
}
\keyword{internal}
