% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{process_kobo_canonised}
\alias{process_kobo_canonised}
\title{Pipeline de canonisation des données Kobo}
\usage{
process_kobo_canonised(kobo, cor, cor_canonique)
}
\arguments{
\item{kobo}{`data.frame`, données Kobo au format long}

\item{cor}{`data.frame`, table de correspondance}

\item{cor_canonique}{`data.frame`, table canonique}
}
\value{
Un `data.frame` final prêt à être exploité
}
\description{
Applique l'ensemble des étapes de transformation et de jointure pour produire
un jeu de données Kobo canonisé.
}
