% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_reseaux}
\alias{process_reseaux}
\title{Calcul des réseaux de mares}
\usage{
process_reseaux(
  df,
  departement,
  use_OS_for_reseaux,
  use_RD_for_reseaux,
  buffer_size_for_reseaux
)
}
\arguments{
\item{df}{Objet sf contenant les données de mares.}

\item{departement}{Code du département.}

\item{use_OS_for_reseaux}{Booléen d'utilisation de l'occupation du sol.}

\item{use_RD_for_reseaux}{Booléen d'utilisation des routes départementales.}

\item{buffer_size_for_reseaux}{Distance tampon utilisée pour les connexions.}
}
\value{
Une liste contenant les réseaux calculés et les mares
enrichies de leur identifiant réseau.
}
\description{
Construit les réseaux de mares et attribue un identifiant
de réseau à chaque mare.
}
\keyword{internal}
