% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{process_traitement_geom}
\alias{process_traitement_geom}
\title{Traitements géométriques IECMAR}
\usage{
process_traitement_geom(sf_kobo, eau, forets, routes)
}
\arguments{
\item{sf_kobo}{`sf`, données géographiques des mares issues de KoboToolbox}

\item{eau}{`sf`, données des surfaces en eau (par défaut `marecoPckg::eau_25`)}

\item{forets}{`sf`, données des zones forestières (par défaut `marecoPckg::forets`)}

\item{routes}{`sf`, données du réseau routier (par défaut `marecoPckg::routes_25`)}
}
\value{
Un `data.frame` contenant les résultats des indicateurs IECMAR
}
\description{
Applique une série de traitements géométriques sur des données de mares afin de
calculer plusieurs indicateurs IECMAR basés sur les distances et les voisinages spatiaux.
}
