% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes}
\alias{routes}
\title{Linéaire routier d'apres la BD TOPO de l'IGN}
\format{
An object of class \code{list} of length 8.
}
\source{
Données originales provenant de la BD TOPO IGN (2025).
}
\usage{
routes
}
\description{
Jeu de donnees des routes format `sf` sur la region Bourgogne-Franche-Comte.
Les geometries ont ete simplifees.
Il est utilisé dans le cadre des traitements géométriques pour le calcul du IECMAr.
}
\keyword{datasets}
