% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_RnAu}
\alias{routes_RnAu}
\title{Linéaire routier d'apres la BD TOPO de l'IGN, filtré sur les RD, RN et autoroutes}
\format{
An object of class \code{list} of length 8.
}
\source{
Données originales provenant de la BD TOPO IGN (2025).
}
\usage{
routes_RnAu
}
\description{
Jeu de donnees des routes format `sf` sur la region Bourgogne-Franche-Comte, filtré sur les RD, RN et autoroutes.
Les geometries ont ete simplifees.
Il est utilisé dans le cadre du calcul des réseaux de mares.
}
\keyword{datasets}
