% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{self_minimal_distance}
\alias{self_minimal_distance}
\title{Calcul de la distance minimale entre entités en excluant les doublons géométriques}
\usage{
self_minimal_distance(sf_object)
}
\arguments{
\item{sf_object}{`sf`, objet spatial contenant une colonne `geom_id`}
}
\value{
Un objet `sf` avec une colonne supplémentaire `dist_mare`
correspondant à la distance minimale à une autre entité distincte
}
\description{
Calcule, pour chaque entité d'un objet `sf`, la distance minimale à une autre
entité en ignorant celles partageant le même identifiant de géométrie (`geom_id`).
}
