% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{split_shp_by_departement}
\alias{split_shp_by_departement}
\title{Sépare les entités d'un shapefile par valeur de la colonne département.}
\usage{
split_shp_by_departement(shp_path, dep_col)
}
\arguments{
\item{shp_path}{Lit un fichier shapefile (sf).}

\item{dep_col}{Colonne où sont disciminer les départements}
}
\value{
Retourne une liste nommée par département, avec chaque élément étant un objet sf.
}
\description{
Utilisée pour la création des JDD. En amont du package (de data-raw -> data.rda).
}
