% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitements_geom.R
\name{st_join_nearest}
\alias{st_join_nearest}
\title{Jointure spatiale sur le plus proche voisin}
\usage{
st_join_nearest(x, y, max_distance = Inf)
}
\arguments{
\item{x}{`sf`, couche spatiale a laquelle on souhaite ajouter les attributs (ex: points)}

\item{y}{`sf`, couche spatiale de reference utilisee pour la jointure (ex: polygones)}

\item{max_distance}{Numeric ou objet `units`, distance maximale en metres pour accepter
une correspondance (default: `Inf`)}
}
\value{
Une couche `sf` contenant les geometries de `x` avec les attributs de `y`
correspondant a la geometrie la plus proche
}
\description{
Associe a chaque geometrie de `x` les attributs de la geometrie la plus proche
dans `y`. Une distance maximale peut etre definie pour limiter les correspondances.
Si aucune geometrie de `y` n'est trouvee dans cette distance, les attributs ajoutes
sont renseignes a `NA`.
}
