% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{try_find_orphelin}
\alias{try_find_orphelin}
\title{Vérification des lignes orphelines après traitement}
\usage{
try_find_orphelin(total, extract1, extract2, extract3)
}
\arguments{
\item{total}{`data.frame`, jeu de données initial au format long}

\item{extract1}{`data.frame`, premier jeu extrait}

\item{extract2}{`data.frame`, deuxième jeu extrait}

\item{extract3}{`data.frame`, troisième jeu extrait}
}
\value{
Un `data.frame` contenant les lignes orphelines (peut être vide)
}
\description{
Identifie les observations du jeu de données initial qui n'ont pas été
retrouvées dans les jeux de données extraits. Permet de détecter d'éventuelles
pertes d'information lors des transformations.
}
