% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{turbidite_if_no_water}
\alias{turbidite_if_no_water}
\title{Compléter la turbidité en absence de donnée (C11)}
\usage{
turbidite_if_no_water(canonique)
}
\arguments{
\item{canonique}{`data.frame`, jeu de données canonisé contenant au moins
les colonnes `X_uuid`, `colname`, `CAN_name`, `CAN_choice`, `id_can`,
et `cor_iecmar`}
}
\value{
Un `data.frame` où les lignes liées à la turbidité sont complétées :
\describe{
  \item{CAN_name}{"turbidite" si absent}
  \item{CAN_choice}{"Inconnu" si valeur manquante}
  \item{cor_iecmar}{Code IECMAR 11 si valeur manquante}
}
}
\description{
Assure qu’une information de turbidité est présente pour chaque mare.
Si aucune valeur n’est renseignée (absence d’eau ou donnée manquante),
une valeur par défaut est attribuée avec le code IECMAR correspondant.
}
\details{
La détection des valeurs manquantes se fait via `id_can`.
Si celui-ci est `NA`, la turbidité est considérée comme non renseignée
et une valeur par défaut est attribuée.
}
