% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validation des paramètres d'entrée}
\usage{
validate_inputs(
  df,
  version,
  departement,
  buffer_size_for_reseaux,
  use_OS_for_reseaux,
  use_RD_for_reseaux,
  use_history_if_exist,
  jdd_v4 = NULL
)
}
\arguments{
\item{df}{Objet sf contenant les données de mares.}

\item{version}{Version du formulaire Kobo utilisée.}

\item{departement}{Code du département.}

\item{buffer_size_for_reseaux}{Distance tampon utilisée pour le calcul des réseaux.}

\item{use_OS_for_reseaux}{Booléen d'utilisation de l'occupation du sol.}

\item{use_RD_for_reseaux}{Booléen d'utilisation des routes départementales.}

\item{use_history_if_exist}{Booléen d'utilisation de l'historique.}

\item{jdd_v4}{Liste des jeux de données V4.}
}
\value{
Invisiblement TRUE si les vérifications sont validées.
}
\description{
Vérifie la cohérence des paramètres transmis à
\code{process_all()} avant exécution des traitements.
}
\keyword{internal}
