% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_iecmar.R
\name{build_gpkg_name2}
\alias{build_gpkg_name2}
\title{Build GeoPackage filename from processing parameters with date/time}
\usage{
build_gpkg_name2(version, departement, prefix = "resultats", dir = "./output")
}
\arguments{
\item{version}{Integer. Processing version used in `process_all()`.}

\item{departement}{Integer or character. Department identifier.}

\item{prefix}{Character. File name prefix. Default is `"resultats"`.}

\item{dir}{Character. Output directory path. Default is `"./output"`.}
}
\value{
Character string. Full path to the generated GeoPackage file.
}
\description{
Build GeoPackage filename from processing parameters with date/time
}
