% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{build_outputs}
\alias{build_outputs}
\title{Construction des sorties IECMAR}
\usage{
build_outputs(compil, res_reseaux, cor_uuid_form_v, df_existe, reseaux)
}
\arguments{
\item{compil}{Tableau compilé des variables IECMAR.}

\item{res_reseaux}{Objet sf contenant les mares et leurs réseaux.}

\item{cor_uuid_form_v}{Correspondance entre identifiants et versions de formulaire.}

\item{df_existe}{Jeu de données filtré sur les mares existantes.}

\item{reseaux}{Réseaux calculés.}
}
\value{
Une liste contenant les différents objets de sortie IECMAR.
}
\description{
Calcule les notes IECMAR et construit les objets de sortie
finaux du traitement.
}
\keyword{internal}
