% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_special.R
\name{default_corridor_v4}
\alias{default_corridor_v4}
\title{Ajouter une valeur par défaut pour le corridor écologique (C14) - V4 uniquement}
\usage{
default_corridor_v4(canonique)
}
\arguments{
\item{canonique}{`data.frame`, jeu de données canonisé contenant au moins
la colonne `X_index`}
}
\value{
Un `data.frame` enrichi avec une ligne par mare :
\describe{
  \item{CAN_name}{"corridor_lineaire_5m"}
  \item{CAN_choice}{"corridor_altere"}
  \item{cor_iecmar}{Code IECMAR C14 (41)}
}
}
\description{
Ajoute, pour chaque mare, une valeur par défaut indiquant un corridor
écologique altéré lorsque l'information n'est pas renseignée dans la V4.
}
\details{
Cette fonction ajoute systématiquement une valeur par défaut pour le corridor
écologique. Elle peut être utilisée lorsque cette information est absente
du jeu de données initial.
}
