% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{defaut_proprietaire_favorable_v4}
\alias{defaut_proprietaire_favorable_v4}
\title{Ajouter un défaut de mesures de protection (C20) pour les formulaires v4}
\usage{
defaut_proprietaire_favorable_v4(df)
}
\arguments{
\item{df}{`data.frame`, contenant au moins la colonne `X_index`}
}
\value{
Un `data.frame` enrichi avec une ligne par mare (si nécessaire) :
\describe{
  \item{CAN_name}{Nom de la variable : "mesures_protection"}
  \item{CAN_choice}{Valeur : "aucune"}
  \item{cor_iecmar}{Code IECMAR C20 (58)}
}
}
\description{
Pour chaque mare, ajoute une ligne indiquant l’absence de mesures de protection
si aucune information n’est déjà présente. Le code IECMAR C20 est fixé à 58.
}
