% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{kobo_wide_to_long}
\alias{kobo_wide_to_long}
\title{Transformation des données Kobo du format large au format long}
\usage{
kobo_wide_to_long(kobo)
}
\arguments{
\item{kobo}{`data.frame`, données Kobo au format large}
}
\value{
Un `data.frame` au format long avec les colonnes :
\describe{
  \item{X_index}{Identifiant de l'observation}
  \item{colname}{Nom de la variable}
  \item{value}{Valeur associée}
  \item{id_temp}{Identifiant technique de ligne}
}
}
\description{
Convertit un jeu de données Kobo du format large au format long,
en conservant un identifiant par ligne et en standardisant les types.
}
