% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_iecmar.R
\name{output_note_only_l}
\alias{output_note_only_l}
\title{Extraire la première note par mare}
\usage{
output_note_only_l(df)
}
\arguments{
\item{df}{`data.frame`, contenant au moins les colonnes :
\code{X_index}, \code{n_critere}, \code{note}}
}
\value{
Un `data.frame` réduit à une ligne par mare avec les colonnes :
\code{X_index}, \code{n_critere}, \code{note}
}
\description{
Sélectionne les colonnes clés liées à la notation IECMAR et conserve
uniquement la première ligne pour chaque mare (`X_index`).
}
